[ISP] Uma interface separada para o hook de renderização
