+++ cotuba/src/main/java/cotuba/md/RenderizadorMDParaHTML.java
@@ -1,6 +1,7 @@
 package cotuba.md;
 
 import cotuba.domain.Capitulo;
+import cotuba.plugin.AoRenderizarHTML;
 import cotuba.plugin.Plugin;
 import org.commonmark.node.AbstractVisitor;
 import org.commonmark.node.Heading;
@@ -80,7 +81,7 @@ public class RenderizadorMDParaHTML {
 
       capitulo.setConteudoHTML(html);
 
-      Plugin.renderizou(capitulo);
+      AoRenderizarHTML.renderizou(capitulo);
 
     } catch (Exception ex) {
       throw new IllegalStateException("Erro ao renderizar para HTML o arquivo " + arquivoMD, ex);
